+++ src/deberta/deberta_model.rs
@@ -732,9 +732,12 @@ impl DebertaForSequenceClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = DebertaConfig::from_file(config_path);
-    /// let model = DebertaForSequenceClassification::new(&p.root(), &config);
+    /// let model = DebertaForSequenceClassification::new(&p.root(), &config).unwrap();
     /// ```
-    pub fn new<'p, P>(p: P, config: &DebertaConfig) -> DebertaForSequenceClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &DebertaConfig,
+    ) -> Result<DebertaForSequenceClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -751,7 +754,11 @@ impl DebertaForSequenceClassification {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
 
         let classifier = nn::linear(
@@ -761,12 +768,12 @@ impl DebertaForSequenceClassification {
             Default::default(),
         );
 
-        DebertaForSequenceClassification {
+        Ok(DebertaForSequenceClassification {
             deberta,
             pooler,
             classifier,
             dropout,
-        }
+        })
     }
 
     /// Forward pass through the model