+++ src/longformer/longformer_model.rs
@@ -791,7 +791,10 @@ pub struct LongformerClassificationHead {
 }
 
 impl LongformerClassificationHead {
-    pub fn new<'p, P>(p: P, config: &LongformerConfig) -> LongformerClassificationHead
+    pub fn new<'p, P>(
+        p: P,
+        config: &LongformerConfig,
+    ) -> Result<LongformerClassificationHead, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
@@ -808,7 +811,11 @@ impl LongformerClassificationHead {
         let num_labels = config
             .id2label
             .as_ref()
-            .expect("num_labels not provided in configuration")
+            .ok_or_else(|| {
+                RustBertError::InvalidConfigurationError(
+                    "num_labels not provided in configuration".to_string(),
+                )
+            })?
             .len() as i64;
         let out_proj = nn::linear(
             p / "out_proj",
@@ -817,11 +824,11 @@ impl LongformerClassificationHead {
             Default::default(),
         );
 
-        LongformerClassificationHead {
+        Ok(LongformerClassificationHead {
             dense,
             dropout,
             out_proj,
-        }
+        })
     }
 
     pub fn forward_t(&self, hidden_states: &Tensor, train: bool) -> Tensor {
@@ -865,21 +872,24 @@ impl LongformerForSequenceClassification {
     /// let device = Device::Cpu;
     /// let p = nn::VarStore::new(device);
     /// let config = LongformerConfig::from_file(config_path);
-    /// let longformer_model = LongformerForSequenceClassification::new(&p.root(), &config);
+    /// let longformer_model = LongformerForSequenceClassification::new(&p.root(), &config).unwrap();
     /// ```
-    pub fn new<'p, P>(p: P, config: &LongformerConfig) -> LongformerForSequenceClassification
+    pub fn new<'p, P>(
+        p: P,
+        config: &LongformerConfig,
+    ) -> Result<LongformerForSequenceClassification, RustBertError>
     where
         P: Borrow<nn::Path<'p>>,
     {
         let p = p.borrow();
 
         let longformer = LongformerModel::new(p / "longformer", config, false);
-        let classifier = LongformerClassificationHead::new(p / "classifier", config);
+        let classifier = LongformerClassificationHead::new(p / "classifier", config)?;
 
-        LongformerForSequenceClassification {
+        Ok(LongformerForSequenceClassification {
             longformer,
             classifier,
-        }
+        })
     }
 
     /// Forward pass through the model